+++ packages/docs/src/stories/Toast.stories.tsx
@@ -1,0 +1,30 @@
+import { Box, Button, Toast, ToastProps } from '@chain-reaction-ui/react'
+import type { StoryObj, Meta } from '@storybook/react'
+import { useState } from 'react'
+
+export default {
+  title: 'Form/Toast',
+  component: Toast,
+  decorators: [
+    (Story) => {
+      const [open, setOpen] = useState(false)
+      return (
+        <Box
+          as={'label'}
+          css={{ display: 'flex', flexDirection: 'column', gap: 2 }}
+        >
+          <Button onClick={() => setOpen((state) => !state)}>Push Toast</Button>
+          {Story({
+            args: {
+              open,
+              title: 'Example Title',
+              description: 'Example description',
+            },
+          })}
+        </Box>
+      )
+    },
+  ],
+} as Meta<ToastProps>
+
+export const Primary: StoryObj<ToastProps> = {}