+++ packages/tokens/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @chain-reaction-ui/tokens
 
+## 3.0.0
+
+### Major Changes
+
+- Add components de Tooltip and Toast
+
 ## 2.1.0
 
 ### Minor Changes