+++ packages/ts-config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @chain-reaction-ui/ts-config
 
+## 3.0.0
+
+### Major Changes
+
+- Add components de Tooltip and Toast
+
 ## 2.0.0
 
 ### Major Changes